Initial pass at refactoring the tag editor to transform it into a tag chooser that can be used both from the song editor and advanced search (and perhaps elsewhere).
I pulled out tagChooser into it's own module and rewired to allow for multiple current states (for search we'll want include/exclude rather than just add) as well as a few cosmetic changes.

Need to enable the multiple current states & then see if it actually works for the search control.

Checking this in now because it appears not to regress current functionality and is a clear improvement over the old code wrt factorization.