Improve (hopefully) the "Song Library" label next to the basic search.
After beating my head against this for way too long (including going down the table path), it looks like we can kludge the bootstrap grid semantics by adding the col-sm & sol-sm offset classes to the <H1> and the div that contains the search control, but then overriding width of the H1 to just let it autosize and the left-margin to the search div to just let is size comfortable against the previous element.

I tested in IE, Chrome, and FF as well as multiple device emulators in Chrome.  This feels a bit kludgy to me, but the other solution was buggying me enough that I'm willing to take the risk...