Fix a bug where push/pull wasn't working correctly on the dance details page.  I ran into this while playing with the site on my laptop.
The basic problem was that there was one div with a col-sm-push-8 and two divs with a col-sm-pull-8.  Mos of the time this worked (I think coincidentally), but sometimes the second div with the pulls (the tempi one) would end up pulled too far to the left.  I saw this in both IE and Chrome.

Fix is to put description and tempi into the same col div...