Added in the wedding dance item to the main page carousel.
In the process of doing this I ran into the evil error that we've been seeing with package restore doings something funky.  This changeset includes the chnages to move from msbuild based to 'automatic' package restore: http://docs.nuget.org/consume/package-restore I'm crossing my fingers that this will be the end of that.

I would prefer to check these things in independently, but I screwed up and didn't shelve my Wedding Dance changes before digging into the nuget changes, both of which hit the m4d.csproj file in a way that's difficult to extricate.