Fix reload tag bug:
This turned out to be because we were deleting tagtypes out from under ourselves rather than a straight up case sensitivity issue.

The case that was firing was Opera:Music vs. opera:Music where the tag type and the first song reference were Opera, but the first song reference was an add then a delete, so it removed the Tagtype, but then attempted to re-add it before updating the database.  Grunt.