Get remove dance completely from song - the first moderator feature - working
This was more painful than expected.  Added the concept of a UserProxy, since we'll be deleting tags for other users.  But otherwise, this is a matter of some ux on the front end, adding a ^ prefix (fully delete) to dance tags and a delete phase to the tag editting funcitonality on the server.