Merged PR 158: Set up edited property
Edited is like modified, but only for actual user edits (not bots/scrapers)

In order to do this I had to fix a bunch of cascading issues
- Load StatsManager at startup and clean up a bunch of stuff around that
- Make ApplicationUser available through more of the system
- Append |P for pseudo user to user fields that are pseudo
- Add a new field to the DB for edited

In addition, during validation I realized that I had introduced a major issue into the spotify connector where I was doing multiple lookups on each song in a playlist even if it had already been added.  Fix this to just fully load the track if it wasn't already part of the playlist.