Merged PR 343: Implement a simpler merge algorithm
Use the new song editor.
Send the client the song histories to merge and a songId for the new song.
The client does the merge, including reconciling album info.
On, save posts to the server.
The server detects the merge state and adds the new song as well as deleting the old ones